# adapter/adapter.rest/adapter.rest.v2/src/main/webapp/api-docs/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "https://github.com/learning-layers/SocialSemanticServer/",
    "version" : "v2",
    "title" : "Social Semantic Server REST API",
    "termsOfService" : "https://github.com/learning-layers/",
    "contact" : {
      "email" : "dtheiler#tugraz$at"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "localhost:8080",
  "basePath" : "/sss.adapter.rest.v2/rest",
...
  "paths" : {
...
    "/circles/circles/filtered/users/{forUser}" : {
      "post" : {
        "tags" : [ "circles" ],
        "summary" : "retrieve filtered circles for user",
        "description" : "",
        "operationId" : "circlesFilteredForUserGet",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "forUser",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSCirclesGetRet"
            }
          }
        }
      }
    },
    "/circles/circles/filtered/{circle}" : {
      "post" : {
        "tags" : [ "circles" ],
        "summary" : "retrieve a filtered circle",
        "description" : "",
        "operationId" : "circleFilteredGet",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "circle",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSCircleGetRet"
            }
          }
        }
      }
    },
    "/circles/circles/users/{forUser}" : {
      "get" : {
        "tags" : [ "circles" ],
        "summary" : "retrieve circles for user",
        "description" : "",
        "operationId" : "circlesForUsersGet",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "forUser",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSCirclesGetRet"
            }
          }
        }
      }
    },
    "/circles/circles/{circle}" : {
      "get" : {
        "tags" : [ "circles" ],
        "summary" : "retrieve a circle",
        "description" : "",
        "operationId" : "circleGet",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "circle",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSCircleGetRet"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "circles" ],
        "summary" : "remove circle",
        "description" : "",
        "operationId" : "circleRemove",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "circle",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSCircleRemoveRet"
            }
          }
        }
      }
    },
    "/circles/circles/{circle}/entities/{entities}" : {
      "post" : {
        "tags" : [ "circles" ],
        "summary" : "add given entities to a circle",
        "description" : "",
        "operationId" : "circleEntitiesAdd",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "circle",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "entities",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSCircleEntitiesAddRet"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "circles" ],
        "summary" : "remove given entities from circle",
        "description" : "",
        "operationId" : "circleEntitiesRemove",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "circle",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "entities",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSCircleEntitiesRemoveRet"
            }
          }
        }
      }
    },
    "/circles/circles/{circle}/type/{type}" : {
      "put" : {
        "tags" : [ "circles" ],
        "summary" : "change circle's type",
        "description" : "",
        "operationId" : "circleTypeChange",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "circle",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "type",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSCircleTypeChangeRet"
            }
          }
        }
      }
    },
    "/circles/circles/{circle}/users/invite/{emails}" : {
      "post" : {
        "tags" : [ "circles" ],
        "summary" : "invite users to a circle",
        "description" : "",
        "operationId" : "circleUsersInvite",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "circle",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "emails",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSCircleUsersInviteRet"
            }
          }
        }
      }
    },
    "/circles/circles/{circle}/users/{users}" : {
      "post" : {
        "tags" : [ "circles" ],
        "summary" : "add given users to a circle",
        "description" : "",
        "operationId" : "circleUsersAdd",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "circle",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "users",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSCircleUsersAddRet"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "circles" ],
        "summary" : "remove users from circle",
        "description" : "",
        "operationId" : "circleUsersRemove",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "circle",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "users",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSCircleUsersRemoveRet"
            }
          }
        }
      }
    },
...
    "/learneps/learneps/circles/{learnEpCircle}" : {
      "put" : {
        "tags" : [ "learneps" ],
        "summary" : "update a circle",
        "description" : "",
        "operationId" : "learnEpVersionCircleUpdate",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "learnEpCircle",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSLearnEpVersionCircleUpdateRet"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "learneps" ],
        "summary" : "remove a circle",
        "description" : "",
        "operationId" : "learnEpVersionCircleRemove",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "learnEpCircle",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/SSLearnEpVersionCircleRemoveRet"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "SSCircleEntitiesAddRet" : {
      "type" : "object",
      "required" : [ "op" ],
      "properties" : {
        "op" : {
          "type" : "string"
        },
        "circle" : {
          "type" : "string"
        }
      }
    },
    "SSCircleEntitiesRemoveRet" : {
      "type" : "object",
      "required" : [ "op" ],
      "properties" : {
        "op" : {
          "type" : "string"
        },
        "entities" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        }
      }
    },
    "SSCircleGetRet" : {
      "type" : "object",
      "required" : [ "op" ],
      "properties" : {
        "op" : {
          "type" : "string"
        },
        "circle" : {
          "$ref" : "#/definitions/SSEntityCircle"
        }
      }
    },
    "SSCircleRemoveRet" : {
      "type" : "object",
      "required" : [ "op" ],
      "properties" : {
        "op" : {
          "type" : "string"
        },
        "circle" : {
          "type" : "string"
        }
      }
    },
    "SSCircleTypeChangeRet" : {
      "type" : "object",
      "required" : [ "op" ],
      "properties" : {
        "op" : {
          "type" : "string"
        },
        "circle" : {
          "type" : "string"
        }
      }
    },
    "SSCircleUsersAddRet" : {
      "type" : "object",
      "required" : [ "op" ],
      "properties" : {
        "op" : {
          "type" : "string"
        },
        "circle" : {
          "type" : "string"
        }
      }
    },
    "SSCircleUsersInviteRet" : {
      "type" : "object",
      "required" : [ "op" ],
      "properties" : {
        "op" : {
          "type" : "string"
        },
        "circle" : {
          "type" : "string"
        }
      }
    },
    "SSCircleUsersRemoveRet" : {
      "type" : "object",
      "required" : [ "op" ],
      "properties" : {
        "op" : {
          "type" : "string"
        },
        "users" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        }
      }
    },
    "SSCirclesGetRet" : {
      "type" : "object",
      "required" : [ "op" ],
      "properties" : {
        "op" : {
          "type" : "string"
        },
        "circles" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/SSEntity"
          }
        }
      }
    },
...
    "SSLearnEpVersionCircleRemoveRet" : {
      "type" : "object",
      "required" : [ "op" ],
      "properties" : {
        "op" : {
          "type" : "string"
        },
        "worked" : {
          "type" : "boolean",
          "default" : false
        }
      }
    },
    "SSLearnEpVersionCircleUpdateRet" : {
      "type" : "object",
      "required" : [ "op" ],
      "properties" : {
        "op" : {
          "type" : "string"
        },
        "worked" : {
          "type" : "boolean",
          "default" : false
        }
      }
    },
...
  }
}